Cleaned up some code, added antialiasing
Cleaned up code and tried randomizing orientation based on signs. I wasn't too impressed with the results of randomizing the signs.

On the other hand, Antialiasing seems to improve the image quality greatly.

I need to upload the three.js-r112 directory in order for this to work on GitHub Pages.