Envmap lighting
This branch applies environment lighting to the snowflakes. It uses envMapIntensity to increase the intensity of the light that the environment casts onto the objects.

I'm not sure if the effect is as good as I was hoping it would be. It seems to make the brightness of the snowflakes more uniform while making their textures harder to see.